Move shared profiler code to profiler_utils.dart
