+++ packages/devtools_app/lib/src/screens/profiler/cpu_profile_transformer.dart
@@ -154,84 +154,6 @@ class CpuProfileTransformer {
   }
 }
 
-// TODO(kenz): move this into a shared file for use with the TimelineEvent
-// class.
-/// Process for converting a [ProfilableDataMixin] into a bottom-up
-/// representation of the profile.
-class BottomUpTransformer<T extends ProfilableDataMixin<T>> {
-  List<T> bottomUpRootsFor({
-    required T topDownRoot,
-    required void Function(List<T>) mergeSamples,
-  }) {
-    final bottomUpRoots = generateBottomUpRoots(
-      node: topDownRoot,
-      currentBottomUpRoot: null,
-      bottomUpRoots: <T>[],
-    );
-
-    // Set the bottom up sample counts for each sample.
-    bottomUpRoots.forEach(cascadeSampleCounts);
-
-    // Merge samples when possible starting at the root (the leaf node of the
-    // original sample).
-    mergeSamples(bottomUpRoots);
-
-    return bottomUpRoots;
-  }
-
-  /// Returns the roots for a bottom up representation of a
-  /// [ProfilableDataMixin] node.
-  ///
-  /// Each root is a leaf from the original [ProfilableDataMixin] tree, and its
-  /// children will be the reverse stack of the original profile sample. The
-  /// stack returned will not be merged to combine common roots, and the sample
-  /// counts will not reflect the bottom up sample counts. These steps will
-  /// occur later in the bottom-up conversion process.
-  @visibleForTesting
-  List<T> generateBottomUpRoots({
-    required T node,
-    required T? currentBottomUpRoot,
-    required List<T> bottomUpRoots,
-  }) {
-    final copy = node.shallowCopy() as T;
-
-    if (currentBottomUpRoot != null) {
-      copy.addChild(currentBottomUpRoot.deepCopy());
-    }
-
-    // [copy] is the new root of the bottom up stack.
-    currentBottomUpRoot = copy;
-
-    if (node.exclusiveSampleCount > 0) {
-      // This node is a leaf node, meaning it is a bottom up root.
-      bottomUpRoots.add(currentBottomUpRoot);
-    }
-    for (final child in node.children.cast<T>()) {
-      generateBottomUpRoots(
-        node: child,
-        currentBottomUpRoot: currentBottomUpRoot,
-        bottomUpRoots: bottomUpRoots,
-      );
-    }
-    return bottomUpRoots;
-  }
-
-  /// Sets sample counts of [node] and all children to [exclusiveSampleCount].
-  ///
-  /// This is necessary for the transformation of a [ProfilableDataMixin] to its
-  /// bottom-up representation. This is an intermediate step between
-  /// [generateBottomUpRoots] and the [mergeSamples] callback passed to
-  /// [bottomUpRootsFor].
-  @visibleForTesting
-  void cascadeSampleCounts(T node) {
-    node.inclusiveSampleCount = node.exclusiveSampleCount;
-    for (final child in node.children.cast<T>()) {
-      child.exclusiveSampleCount = node.exclusiveSampleCount;
-      cascadeSampleCounts(child);
-    }
-  }
-}
-
 /// Merges CPU profile roots that share a common call stack (starting at the
 /// root).
 ///
